refactor: Post 삭제 기능 분리
